move internals around to make way for NEL and reporting API
Updates the internals to be better organised for when we introduce support for NEL and reporting API endpoints.
