+++ go.mod
@@ -1,7 +1,10 @@
 module github.com/jacobbednarz/go-csp-collector
 
-go 1.18
+go 1.19
 
-require github.com/sirupsen/logrus v1.9.0
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/sirupsen/logrus v1.9.0
+)
 
 require golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
