+++ internal/handler/healthcheck.go
@@ -1,0 +1,9 @@
+package handler
+
+import (
+	"net/http"
+)
+
+func HealthcheckHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+}
