+++ internal/handler/notfound.go
@@ -1,0 +1,9 @@
+package handler
+
+import (
+	"net/http"
+)
+
+func NotFoundHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusNotFound)
+}
